move more functions to cluster builder
